add values to userSettings, add Promise reflect
username, email and language are now available in userSettings.
dashboard pre-rendering is now done using bluebird's Promise.reflect because Promise.all would crash the process with a single error and nothing would get rendered.
for this to work, I had to implement bluebird on the Promises returned in the /models.